UPDATE: Barbedo. Several bug-fixes, logging and documenting. FIX: SpecRollOff. Calculation should be done the right now. FIX: WAVHandler. Fix a rounding problem that was making the result conflict with matlab. UPDATE: Hamming. House keeping.

+++ src/Extensions/Handlers/WAVHandler/WAVHandler.cs
@@ -173,7 +173,8 @@ namespace MusiC.Extensions.Handlers
                             
                             // Increases the number of divisions but avoid overflow problems
                             // Makes data mono
-                            *pData += temp / _channels;
+                            // long / int = int .... need to cast to float.
+                            *pData += (float) temp / _channels;
                         }
 
                         pData++;
